requirements.txt
# Core dependencies
# NOTE: The codebase uses the OpenAI Whisper API via `import whisper`.
# Prefer the official openai-whisper package. The timestamped fork is commented out for now.
# whisper-timestamped>=1.12.0
openai-whisper>=20231117
torch>=2.0.0
sentence-transformers>=2.2.2
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Open-source LLM (replace OpenAI)
transformers>=4.30.0
accelerate>=0.20.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-benchmark>=4.0.0 